Add an mcp extra to dagster-dg
Summary:
As we likely move away from dagster-dg being only installed in an isolated global environment, we have to be a bit stricter about what we bring in. In that light, make mcp a separate extra while it's still an experimental feature.

`dagster mcp serve` with and without the relevant extra installed

> Insert changelog entry or delete this section.